requirements.txt
streamlit
pandas
scikit-learn
matplotlib
biopython
